[8bcf664] [cleanup] moves IGUIContainer into a separate include and drops the GUIBaseContainer.h include from GUIViewControl.h.
