+++ xbmc/GUIViewControl.h
@@ -20,9 +20,12 @@
  *
  */
 
-#include "GUIViewState.h"
+#include <vector>
+#include "utils/StdString.h"
+#include "guilib/GraphicContext.h" // for VIEW_TYPE
 
-#include "GUIBaseContainer.h"
+class CGUIControl;
+class CFileItemList;
 
 class CGUIViewControl
 {
