+++ xbmc/Util.cpp
@@ -63,6 +63,8 @@
 #include "music/dialogs/GUIDialogMusicScan.h"
 #include "dialogs/GUIDialogFileBrowser.h"
 #include "video/dialogs/GUIDialogVideoScan.h"
+#include "guilib/TextureManager.h"
+#include "guilib/IGUIContainer.h"
 #include "utils/fstrcmp.h"
 #include "utils/Trainer.h"
 #ifdef HAS_XBOX_HARDWARE
@@ -1582,31 +1584,31 @@ void CUtil::RemoveIllegalChars( CStdString& strText)
 {
   char szRemoveIllegal [1024];
   strcpy(szRemoveIllegal , strText.c_str());
-  static char legalChars[] = " ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz1234567890!#$%&'()-@[]^_`{}~.ßåÄäÖöüøéèçàùêÂñáïëìíâãæîğòôóõ÷ú";
+  static char legalChars[] = " ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz1234567890!#$%&'()-@[]^_`{}~.ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½";
   
   char *cursor;
   for (cursor = szRemoveIllegal; *(cursor += strspn(cursor, legalChars)); /**/ )
   {
     // Convert FatX illegal characters, if possible, to the closest "looking" character:
-    if (strchr("ÂÁÀÄÃÅ", (int) *cursor)) *cursor = 'A';
+    if (strchr("ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½", (int) *cursor)) *cursor = 'A';
     else
-    if (strchr("âáàäãå", (int) *cursor)) *cursor = 'a';
+    if (strchr("ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½", (int) *cursor)) *cursor = 'a';
     else
-    if (strchr("ÔÓÒÖÕ", (int) *cursor)) *cursor = 'O';
+    if (strchr("ï¿½ï¿½ï¿½ï¿½ï¿½", (int) *cursor)) *cursor = 'O';
     else
-    if (strchr("ôóòöõ", (int) *cursor)) *cursor = 'o';
+    if (strchr("ï¿½ï¿½ï¿½ï¿½ï¿½", (int) *cursor)) *cursor = 'o';
     else
-    if (strchr("ÛÚÙÜ", (int) *cursor)) *cursor = 'U';
+    if (strchr("ï¿½ï¿½ï¿½ï¿½", (int) *cursor)) *cursor = 'U';
     else
-    if (strchr("ûúùüµ", (int) *cursor)) *cursor = 'u';
+    if (strchr("ï¿½ï¿½ï¿½ï¿½ï¿½", (int) *cursor)) *cursor = 'u';
     else
-    if (strchr("ÊÉÈË", (int) *cursor)) *cursor = 'E';
+    if (strchr("ï¿½ï¿½ï¿½ï¿½", (int) *cursor)) *cursor = 'E';
     else
-    if (strchr("êéèë", (int) *cursor)) *cursor = 'e';
+    if (strchr("ï¿½ï¿½ï¿½ï¿½", (int) *cursor)) *cursor = 'e';
     else
-    if (strchr("ÎÍÌÏ", (int) *cursor)) *cursor = 'I';
+    if (strchr("ï¿½ï¿½ï¿½ï¿½", (int) *cursor)) *cursor = 'I';
     else
-    if (strchr("îìíï", (int) *cursor)) *cursor = 'i';
+    if (strchr("ï¿½ï¿½ï¿½ï¿½", (int) *cursor)) *cursor = 'i';
     else
     *cursor = '_';
   }
