+++ xbmc/guilib/GUIBaseContainer.h
@@ -25,12 +25,9 @@
  *
  */
 
-#include "GUIControl.h"
+#include "IGUIContainer.h"
 #include "GUIListItemLayout.h"
-#include "boost/shared_ptr.hpp"
 #include "utils/Stopwatch.h"
-
-typedef boost::shared_ptr<CGUIListItem> CGUIListItemPtr;
 
 /*!
  \ingroup controls
@@ -38,23 +35,6 @@ typedef boost::shared_ptr<CGUIListItem> CGUIListItemPtr;
  */
 
 class IListProvider;
-
-class IGUIContainer : public CGUIControl
-{
-protected:
-  VIEW_TYPE m_type;
-  CStdString m_label;
-public:
-  IGUIContainer(int parentID, int controlID, float posX, float posY, float width, float height);
-  virtual bool IsContainer() const { return true; };
-
-  VIEW_TYPE GetType() const { return m_type; };
-  const CStdString &GetLabel() const { return m_label; };
-  void SetType(VIEW_TYPE type, const CStdString &label);
-
-  virtual CGUIListItemPtr GetListItem(int offset, unsigned int flag = 0) const = 0;
-  virtual CStdString GetLabel(int info) const                                  = 0;
-};
 
 class CGUIBaseContainer : public IGUIContainer
 {
